correcion

+++ my-app/src/main/java/com/mycompany/app/DTO/CursoDTO.java
@@ -1,12 +1,12 @@
 package com.mycompany.app.DTO;
 
 public class CursoDTO {
-  private Integer ID;
+  private Double ID;
   private String nombre;
   private ProgramaDTO programa;
   private Boolean activo;
 
-  public CursoDTO(Integer ID, String nombre, ProgramaDTO programa, Boolean activo) {
+  public CursoDTO(Double ID, String nombre, ProgramaDTO programa, Boolean activo) {
     this.ID = ID;
     this.nombre = nombre;
 
@@ -14,7 +14,7 @@ public class CursoDTO {
     this.activo = activo;
   }
 
-  public Integer getID() {
+  public Double getID() {
     return ID;
   }
